Создаёт eventController - переносит логику отрисовки событий, точек маршрута в контроллер
